+++ cmd/vroom/config.go
@@ -2,8 +2,9 @@ package main
 
 type (
 	ServiceConfig struct {
-		Environment string `env:"SENTRY_ENVIRONMENT" env-default:"development"`
-		Port        int    `env:"PORT"               env-default:"8085"`
+		Environment    string `env:"SENTRY_ENVIRONMENT" env-default:"development"`
+		Port           int    `env:"PORT"               env-default:"8085"`
+		WorkerPoolSize int    `env:"WORKER_POOL_SIZE"               env-default:"100"`
 
 		SentryDSN string `env:"SENTRY_DSN"`